pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rl-volatility-alpha"
version = "0.1.0"
description = "RL-Driven Volatility Forecasting and Allocation Using Market Microstructure & Unstructured Data"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "DataHiveMind", email = "your.email@example.com"}
]
keywords = [
    "reinforcement-learning",
    "volatility-forecasting",
    "market-microstructure",
    "options-trading",
    "quantitative-finance",
    "deep-learning",
    "pytorch"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    # Core ML/DL frameworks
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    
    # RL libraries
    "stable-baselines3>=2.0.0",
    "sb3-contrib>=2.0.0",
    "gymnasium>=0.28.0",
    
    # Data processing
    "pandas>=2.0.0",
    "polars>=0.19.0",
    "pyarrow>=12.0.0",
    "arcticdb>=4.0.0",
    
    # Financial data & analysis
    "yfinance>=0.2.0",
    "ta-lib>=0.4.0",
    "arch>=6.0.0",
    "quantlib>=1.31",
    
    # Time series
    "statsmodels>=0.14.0",
    "pmdarima>=2.0.0",
    
    # NLP & transformers (for unstructured data)
    "transformers>=4.30.0",
    "sentence-transformers>=2.2.0",
    "tokenizers>=0.13.0",
    
    # Experiment tracking & MLOps
    "mlflow>=2.8.0",
    "tensorboard>=2.13.0",
    "wandb>=0.15.0",
    
    # Configuration & utilities
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.0",
    
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.14.0",
    
    # Performance & optimization
    "numba>=0.57.0",
    "cython>=3.0.0",
    
    # Testing & quality
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    
    # Jupyter
    "jupyter>=1.0.0",
    "ipykernel>=6.23.0",
    "ipywidgets>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    "pre-commit>=3.3.0",
    "ipdb>=0.13.13",
]

docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "sphinx-autodoc-typehints>=1.23.0",
    "myst-parser>=2.0.0",
]

all = [
    "rl-volatility-alpha[dev,docs]",
]

[project.urls]
Homepage = "https://github.com/DataHiveMind/Reinforcement_Learning_for_Volatility"
Documentation = "https://github.com/DataHiveMind/Reinforcement_Learning_for_Volatility/blob/main/README.md"
Repository = "https://github.com/DataHiveMind/Reinforcement_Learning_for_Volatility"
"Bug Tracker" = "https://github.com/DataHiveMind/Reinforcement_Learning_for_Volatility/issues"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
src = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | experiments
  | data
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["**/experiments/*", "**/data/*"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "stable_baselines3.*",
    "gymnasium.*",
    "arcticdb.*",
    "mlflow.*",
    "transformers.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = """
    -v
    --strict-markers
    --cov=src
    --cov-report=term-missing
    --cov-report=html
    --cov-report=xml
"""
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/experiments/*",
    "*/notebooks/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
